requirements.txt
streamlit
streamlit-folium
Pillow
geopy
folium
piexif
